backend/requirements.txt
Flask
langchain
langchain-openai
langchain-community
serpapi
langchain-tavily
python-dotenv
requests
PyMySQL
cryptography
flask_jwt_extended 